requirements.txt
certifi==2018.10.15
chardet==3.0.4
cryptography==2.1.4
Django==2.1.3
httplib2==0.9.2
lazy-object-proxy==1.3.1
macaroonbakery==1.1.3
Mako==1.0.7
MarkupSafe==1.0
mccabe==0.6.1
numpy==1.15.4
oauth==1.0.1
olefile==0.45.1
pandas==0.23.4
psycopg2==2.7.6
pycrypto==2.6.1
requests==2.20.1
urllib3==1.24.1
virtualenv==15.1.0
django-leaflet == 0.24.0
django-improved-user == 1.0.0
django-geojson == 2.11.0
